feat: 매주 화요일 등위별 당첨자 선별 schedule 로직, 사용 회차별 당첨정보 요청

+++ src/main/java/lotto/practice/random/domain/winningAmount/rank/TotalCountFactory.java
@@ -2,7 +2,7 @@ package lotto.practice.random.domain.winningAmount.rank;
 
 public class TotalCountFactory {
 
-    public static TotalCount setRankTotalCount(int fifthTotalCount, int fourthTotalCount, int thirdTotalCount, int secondTotalCount, int firstTotalCount) {
+    public static TotalCount setRankTotalCount(Long fifthTotalCount, Long fourthTotalCount, Long thirdTotalCount, Long secondTotalCount, Long firstTotalCount) {
         return TotalCount.builder()
                 .fifthTotalCount(fifthTotalCount)
                 .fourthTotalCount(fourthTotalCount)
