+++ src/main/java/lotto/practice/random/domain/winningInfo/RankAmountRepository.java
@@ -4,7 +4,7 @@ import lotto.practice.random.domain.Winner.RankType;
 
 public interface RankAmountRepository {
 
-    int getRankTotalCount(RankType rank, Long lottoCycleNum);
+    Long getRankTotalCount(RankType rank, Long lottoCycleNum);
 
     Long getTotalSellAmount(Long lottoCycleNum);
